원성분토크나이저,파일정리
