--- crawling/유사어집합/분류.txt
@@ -1,429 +1,0 @@
-냉가리,돝고기,돼지고기,박나물,돈육,제육
-소맥분,진말,밀가루,소맥면,맥분
-해바라기씨유,해바라기유,해바라기기름
-소금,소조
-냉우,처우,찬비,한우
-바닷소금,볕소금,천일염
-구연산,시트르산,레몬산
-쇠갈비,소갈비
-식물유지,식물성오일,식물성기름,식물성유지,식물유
-두유,태유,대두기름,대두유,콩기름
-실리콘수지,실리콘,규소수지
-계란,계단,계자,달걀
-대산,코끼리마늘,호산,마늘
-흰쌀,백미
-알타리,총각무
-타피오카,타피오카전분
-종균,씨균
-설탕,가루사탕,사탕,사탕가루
-계육,닭고기
-두포,두부
-당근,당빠
-호초,후추
-홍도,경파,홍파,경랑,거도,대파,경도
-겨기름,쌀겨유,미강유,쌀겨기름
-물엿,수이
-미곡,쌀,대미
-토코페롤,비타민이
-글루코오스,글루코스,포도당,당글루코오스
-임금산,사과산,능금산,말산
-글루탐산나트륨,글루탐산소다
-글라이신,글리신
-미분,쌀가루
-백설탕,백당
-락토오스,젖당,유당,락토스
-타락,쇠젖,소젖,우유
-일년감,번가,남만시,토마토
-유청분말,유청가루
-포스파티딜콜린,레시틴,콜린인산유도체
-우육,왕눈사지,소고기,쇠고기,황육
-소르비톨,소비톨,글루시톨,소르바이트,소바이트
-토장,된장
-파인애플,봉리수
-양육,양고기
-검정콩,흑콩,흑대두,검은콩
-연근,연뿌리,연우,우근
-감사,고추장
-소맥글루텐,밀글루텐
-적주,적포도주
-표고,표고버섯
-참치부제,참치
-감,단과
-증기식공기추출기,에스제이,증기이젝터,증기분사식공기추출기
-원당,원료당
-종국,만미
-저지,돈지,돼지기름,돼지비계
-선척,배,선박,주선
-묵은 김치,묵은지
-나미,찹쌀,점갱미,점미
-콩나물,숙아채,두아
-시금치,파릉채,마아초,적근채
-소두,팥
-무염버터,무가염버터
-옥수수,옥수수나무
-평과,빈파,사과
-세겹살,삼겹살
-멸어,멸아,행어,추어,멸치,이준,말자어
-유산석회,황산칼슘,유산칼슘
-다시마,해대,참다시마,곤포
-차돌박이살,차돌박이
-홍합,담채,이패,해폐,담치,참담치
-국간장,조선간장
-법유,들기름
-욱키,숨,물,임계,땅소주,산화수소,우케,우끼미,우크미,수음,임공
-덱스트린,호정
-괵실,도토리,곡실
-올방개,지율
-녹두채,숙주,숙주채,숙주나물
-궐채,고사리
-참기름,호마유,진유,지마유,향유
-대순,죽순,죽태
-담양,담양읍
-오른발,우족,바른발
-숙변,숙지황,숙하
-사선목,두충
-석조,산수유
-농축훼이단백,농축유청
-레닌,키모신,응유효소
-무수아류산,이산화황,아황산가스,이산화유황,아황산와사,무수아황산,아황산무수물
-채종유,채유
-옥수수배아,옥배아박
-농축대두단백질,농축대두단백
-노른자,노른자위,난황,노른자알덩어리
-흑유마,흑지마,흑호마,검정깨,거승,검은깨
-야자기름,야자유
-글리세린,감유,글리세롤
-앞다리,전지
-현미,반미,반도미,조미,매조미쌀
-우방,우엉
-당면,분탕,호면
-자일로오스,크실로오스,목재당,목당
-초,식초
-이스트,발효균,효모균,효모
-신안군,신안
-마그네슘클로라이드,염화마그네슘
-식용유,식유,식료유
-아몬드,감편도
-대조,목밀,대추
-대두,대적,저당,저적
-오구,인릉,문동,계전초,도미,양구,우구,맥문동,애구,불사초
-분홍,무지개,천궁
-초산나트륨,아세트산나트륨,초산소듐,초산소다
-지정,인삼
-아르지닌,아르기닌
-콜라겐,아교질,교원질
-젖산박테리아,유산균,젖산세균,젖산균
-흑미,검은 쌀,흑색미,검정 쌀
-메밀,목맥,교맥,오맥
-비후비염,비후코,비후성비염
-산성탄산나트륨,탄산수소나트륨,중조,중탄산소다,중탄산나트륨
-백회,석회,탄산석회,탄산칼슘
-관계관리,RM,거래처종합관리
-다이오드트랜지스터논리회로,다이오드트랜지스터로직,다이오드트랜지스터논리,DTL
-유청,젖국
-백유마,진임,백지마,백호마,참깨
-우자,타로감자,토지,토란,토련,땅토란
-CMC,컴퓨터뮤직센터
-흑사탕,흑설탕,흑당
-암모늄황산염,황산암모니아,유안,유산암모늄,황산암모늄
-유용우,유우,착유우,젖소
-느타리버섯,느타리
-소내장,작은내장신경
-인산이수소칼슘,제일인산칼슘
-생강,생,새앙
-국소충수염,소막창
-단백효소,프로테아제,단백질가수분해효소
-식품첨가제,식품첨가물
-돼지족발,족발
-탄산음료,탄산,청량음료
-탄산무수물,무수탄산,탄산가스,이산화탄소
-스탠더드넘버,에버
-태어,명태어,명태
-모맥,보리,대맥,겉보리
-렌틸콩,렌즈콩
-백양,백태
-꿀,봉밀,청밀,청,백밀,벌꿀
-압육,오리고기
-생강분,새앙가루,생강가루
-신미,매운맛
-국균,노란곰팡이,황국균,누룩곰팡이
-건맥아,엿길금,엿기름,맥아유,맥얼,맥아
-뜸씨,뜸팡이,효소
-김자반,감태자반
-채,폄,편,떡,망짜
-명음반,명반
-꽃소금,재제염,재염
-강도,핵도,당추자,호두
-홍화유,홍화씨유
-지,요강
-가경자,자두,자리
-인산이수소나트륨,제일인산나트륨
-청어,고등어
-착색료,착색제
-할라페뇨,할라피뇨
-크랜베리,덩굴월귤,넌출월귤
-전화당,반전당
-염화칼륨,염화포타슘,염화칼리,클로르칼륨
-등주,등심
-수동식교환,사설수동교환,프라이빗수동교환,피엠엑스
-비,빗자루
-향신료,신향료
-혼합곡,혼합 곡식
-소선,소형선,단주,편주,조각배,소주,일편주,개주,소정,단정,쪽배
-토시살,행잉텐더
-제삼인산칼슘,이인산삼칼슘,삼차인산칼슘
-식물성,식물질
-무릎도가니,도가니
-선지피,선지
-토란대나물,토란줄기
-감자,사탕수수
-코코아,코코아파우더
-콩가루,두황
-땅콩,땡꽁
-말제,분약,산제,분제,분말약,분말제,산,산약,가루약
-오미자나무,오미자
-석창포,석장포
-채용,녹용
-갱탕,국
-인산,인해
-고운대,토란대
-시래기,청경
-포도당주사액,포도당액
-기외,기타,여타
-심장,염통,중단전
-부소,글루텐,맥부질,부질
-탄산칼리,탄산칼륨,칼리
-탄산소다,소다,탄산나트륨
-제이인산나트륨,인산수소나트륨
-생삼,수삼
-겨잣가루,겨자분
-닭,덕금
-연황미,호박고구마
-콜라드,케일
-셀러리,양미나리
-곱창,소창
-누룩,주매,곡자,국자,은국,국얼
-호밀,양맥,라이보리,흑맥
-야임,들깨,임자,백소,수임
-야채,들나물
-쌀농사,벼농사,도작,미작
-강력밀가루,강력분
-산양유,염소젖
-근채,미나리,개미나리,수근
-맑은술,징주,청주
-산화제이주석,주석산,산화제이석,이산화주석
-co,원산지증명서
-지식표현,kr
-목심,나뭇고갱이
-살균우유,살균유
-배양액,배양기
-흑염소,흑산양
-β,아드레날린성베타수용체
-인산이수소포타슘,케이디피,제일인산칼륨
-니코틴산아마이드,니코틴산아미드
-동부,이복,이모,각배,별복
-공어,석려,홍어
-에프디,열교차단장치,방화댐퍼
-난백분,난백소
-시멘트분산제,분산제
-집단구분자,지에스
-열매,과실,과자
-난황분,노른자가루,분말계란
-유지,유지방,젖기름
-단백,난백
-팽이버섯,팽나무버섯
-cw,연속파마그네트론
-중탄산암모니아,탄산수소암모늄
-유정란,수정란
-뜸팡잇가루,건조효모
-트라이아세틴,트리아세틴
-프로판산,프로피온산
-천연색소,자연색소
-락색소,락카
-인조미,인공미
-흰죽,이죽,쌀죽
-돈갈비,돼지갈비
-서양고추,피망
-마카로니밀,듀럼소맥,듀럼밀
-돼지 뼈,돈골
-제,제도
-내로스프레드,레드칼라
-우물물,정수
-벤조산나트륨,안식향산나트륨
-쌀겨,미강
-후지,뒷다리,후각
-당밀,사탕밀
-돈피,돼지가죽
-새송이,새송이버섯
-강낭콩,강남두
-분,분말,가루
-의이,율무
-기장,옷기장
-나속,차조
-쥐치,쥐치어
-제이인산칼륨,인산수소포타슘
-밀보리,쌀보리,과맥,청과맥,나맥
-오골계,실키종
-능이,능혈,향버섯,능이버섯
-소위원회,소위
-맛기름,향미유
-바실대다,바실
-제삼인산나트륨,삼차인산나트륨,인산나트륨
-쇠뼈다귀,소뼈,쇠뼈,우골
-맥아당,말토스,말토오스,엿당
-시트르산철암모늄,구연산철암모늄
-질산나트륨,질산소다
-CV,시중점
-아이오딘칼리,요오드화칼륨,옥화칼륨,요오드칼리,아이오딘화포타슘,옥도칼륨,요오드칼륨,아이오딘화칼륨,옥도가리,아이오딘칼륨,옥화포타슘
-새똥,조분
-유수분화기,수수분
-영자기장갈라지기,영자기마당갈라지기,무자기장갈라지기,무자기마당갈라지기,기장분
-www,세계기상감시계획
-울보,우지
-경화유,굳힌기름
-귀견줌,비이
-카사바녹말,카사바전분
-바지락조개,바지락,참조개,바지라기,바지랑이,합리
-아밀레이스,디아스타아제,아밀라아제,아밀라제,다이아스테이스
-아비시니아,에티오피아
-아스파라진산,아미노호박산,아스파트산,아스파라긴산,아스파르트산
-CPP,가변피치프로펠러
-과일즙,실과즙,열매즙,과일액,과실액,과즙,과실즙
-남아프리카공화국,남아공
-산화아연,아연화,아연백
-등급,제약등급,제약위계,위계
-카르니틴,카니틴,비타민비티
-소양지차,천지 차,소양지판,소양지,천양지차,천지 차이,운니지차
-늙은호박,청둥호박,맷돌 호박
-메밀가루,목말,백면,백면가루,교맥분
-간당,간도
-엿,이당
-인산수소칼슘,제이인산칼슘
-잎산,프테로일글루타민산,비타민엠,프테로일글루탐산,폴산,엽산
-뽕,상엽,뽕잎
-순두부,수두부
-이미터폴로어,이미터추종,EF
-울금,심황,강황
-식빵,식면,식면포
-국물김치,물김치
-소어,밴댕이,반초어
-황태,더덕,더덕북어
-락트산,유산,젖산
-피에프,프로젝트파이낸싱
-가시오가피,가시오갈피
-삽주,산정,마계,걸력가,산계,산강
-배초향,방아
-방약과,방과,모과
-홍차,홍다
-카모마일,캐모마일
-마테,자모기둥
-쥐치포,쥐포
-밀배아박,탈지밀
-핵산,뉴클라인산,누클라인산
-후추분,호초말,후춧가루
-아황산나트륨,아황산소다
-이,이씨
-유자정과,유자청
-어니언,슈퍼레이어드룩,헤비레이어드룩
-아일랜드,에이레
-잡어,잡고기
-채끝,채끝살
-설검,설도
-번데,회용,번데기,귀용
-다리몽둥이,족,다리갱이,하지,다리깽이,각공이,다리,각,족다리,버데기,아래가지
-손바닥선인장,부채선인장,백년초
-과줄,약과
-매자과,타래과,매엽과,매작과,매잡과
-치인,치자
-산딸기나무,산딸기
-북채,경채,베이징요리
-우둔살,우둔,방심,방심살
-사태,사태고기
-결명자,환동자
-교재센터,시청각센터,MRC
-훈조,메주
-상황버섯,상황
-상실,상심,오디
-석류나무,석류
-병,이수,병막,병귀,병증,병마
-에이알고리즘,에이스타
-고채,고들빼기
-제삼인산칼륨,인산포타슘,삼차인산칼륨
-고식책,고식지계,목전지
-정신신경안정제,정온제,안정제,신경안정제,정신안정제,정온약
-로이신,류신
-피에이엘,페닐알라닌
-소브산칼륨,소르브산칼륨
-시트르산칼륨,구연산칼륨
-밀녹말,밀전분
-난황색,노른자색
-암페어시간,암페어시,암페어아워,에이에이치
-콩장,콩간장
-고수,고두
-훈조태,메주콩
-연어반,연어살색반,연어살
-청국장,담북장
-백후추,흰 후추
-염화암모니아,염안,암모늄클로라이드,염화암모늄,염사,염화암몬
-수오미,핀란드
-미료,조미료
-라이머콩,흰강낭콩
-자초,검은담비,흑초,초웅,산달,잘
-쯔란,쿠민
-겨자가스,머스타드,이페리트
-양송이,양송이버섯,머시룸
-옥수숫가루,옥수수분
-엑스클루시브오어게이트,or,엑스오어게이트
-잡꿀,잡화꿀
-토사자,새삼씨
-진자,헤이즐넛,개암
-적색,빨강
-황색,누런색,황,누런빛
-카페인,다소,커피소,다정
-가성칼리,수산화칼륨,수산화포타슘
-이노지트,근육당,이노시톨
-인산칼슘,인산석회
-결합인덱스,바인덱스,결합색인
-바이타민,생활소,비타민
-초산,첫몸풀기
-건조무,무미건조
-장담,장언,장어
-적소두,적두,붉은팥,홍두
-당,망건당
-전분당,녹말당
-국화,나라꽃
-오층위,o
-오가피,오갈피
-백배,쇠무릎,대절채,쇠무릎지기,우슬,산현채
-슈츠스타펠,나치친위대,에스에스
-말랑무,말랭이,무말랭이
-긴맛살,죽합,죽정,맛조개,맛살,구신맛,긴맛,마도패
-제이인산암모늄,인산수소암모늄
-벤즈알데히드,벤즈알데하이드
-지엠식품,지엠푸드,GMO
-도달,당도
-복분자딸기,나무딸기,고무딸기,복분자
-산사자,산사,당구자,아가위,목도
-영지버섯,지초,영지,불로초
-녹다운수출,맞춤식,KD,녹다운,녹다운방식,케이디수출
-우즈베크,우즈베키스탄
-적채,적양배추
-백설고,설기,백편,흰무리,백설기,흰무리떡
-갓,입자
-밀기울,맥피,소맥피,소맥부,맥부,밀찌꺼기,대피
-탁주,막걸리,탁료
-건조채소,건조야채
-건파,절뚝이,건각,절뚝발이,파벽,뚝발이,편파
-차아황산소다,차아황산나트륨,티오황산나트륨,싸이오황산나트륨,티오황산소다,차아유산나트륨,싸이오황산소다
-꼬시래기,강리
-건멸치,마른 멸치
-아목,개두릅나무,해동,음나무,엄나무
-질산염,초산염
-국제상품분류,HS
-핵자기공명,엔엠,핵자기껴울림
-자담,자당
-건새우,마른 새우
-젖산철,락트산철
-섬유소,피브린
-박하뇌,박하상,멘톨,박하빙,박하정
-꽃상추,엔디브
